refactor(fe): address MCP settings review nits
- Dialog owns its own saving state by awaiting the async onSave, so the
  parent no longer threads a saving prop down.
- Drop the unreachable add-connector button and enableOnSave branch:
  enabling always goes through the dialog, which trusts-and-enables in
  one write.
- Generalize the e2e holdToConfirm util to take a button name and wait
  for it to be enabled, replacing the two duplicated local helpers.
