+++ src/frontend/tests/e2e-playwright/utils.ts
@@ -289,13 +289,13 @@ export const openIiTab = async (page: Page): Promise<Page> => {
 // and mobile contexts) — the component's listener fires on any synthetic
 // event. The controller releases itself at the 2500ms duration, so no
 // matching mouseup is needed.
-export const holdToConfirm = async (page: Page): Promise<void> => {
-  await page
-    .getByRole("button", { name: "Hold to confirm" })
-    .dispatchEvent("mousedown");
-  await page
-    .getByRole("heading", { level: 1, name: "Enter the code" })
-    .waitFor();
+export const holdToConfirm = async (
+  page: Page,
+  name = "Hold to confirm",
+): Promise<void> => {
+  const button = page.getByRole("button", { name });
+  await expect(button).toBeEnabled();
+  await button.dispatchEvent("mousedown");
 };
 
 /**
